src/cli/strategies/ISO31661Alpha2Strategy.ts
import { PropertyDeclaration } from "ts-morph";
import inquirer from "inquirer";
import { ValidationStrategy } from "./ValidationStrategy";

export class ISO31661Alpha2Strategy extends ValidationStrategy {
  async execute() {
    const properties = this.classes[0].getProperties();
    const propertyNames = properties.map((property: PropertyDeclaration) =>
      property.getName()
    );

    const answers = await inquirer.prompt([
      {
        name: "property",
        message: "Select the property to add ISO 3166-1 alpha-2 validation:",
        type: "list",
        choices: propertyNames,
      },
      {
        name: "message",
        message: "Validation error message:",
        type: "input",
      },
    ]);

    const property = this.classes[0]
      .getProperties()
      .find((p) => p.getName() === answers.property);
    if (property) {
      property.addDecorator({
        name: "ISO31661Alpha2",
        arguments: [
          `{
            message: "${answers.message}"
          }`,
        ],
      });
    }
  }
}
